refactor: centralize secret patterns and sanitization logic for output scrubbing
